insert working

+++ dbms/dbms/tokenizer.h
@@ -6,13 +6,13 @@
 class Token {
 public:
 	char kind;        // what kind of token
-	double value;     // for numbers: a value 
+	int value;     // for numbers: a value 
 	char letter;
 	Token(char ch)    // make a character Token from a char
 		:kind(ch), value(0), letter('\0') { }
 	Token(char ch, char ch2)	// make a letter Token from a char and a letter
 		:kind(ch), value(0), letter(ch2){ }
-	Token(char ch, double val)   // make a number Token from a char and a double
+	Token(char ch, int val)   // make a number Token from a char and a double
 		:kind(ch), value(val), letter('\0') { }
 
 };
